rename refresh_token.py to hash.py for convenience, delete password.py, change router file name
